refactor: 연관성 있는 파일 별로 패키지 분리
