--- app/src/main/java/woowacourse/movie/domain/model/ReservationInfo.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.domain.model
-
-class ReservationInfo(
-    val reservationCount: Int,
-    seatingChart: SeatingChart,
-) {
-    val selectedSeats = SelectedSeats(seatingChart, reservationCount)
-}
