+++ app/src/main/java/woowacourse/movie/domain/model/reservation/date/ScreeningTime.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.reservation.date
 
 import java.time.LocalTime
