+++ app/src/main/java/woowacourse/movie/domain/model/reservation/seat/Seat.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.reservation.seat
 
 class Seat(
     val row: Int,
