+++ app/src/test/java/woowacourse/movie/domain/model/reservation/date/ScreeningTimeTest.kt
@@ -1,9 +1,10 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.reservation.date
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
+import woowacourse.movie.domain.model.reservation.date.ScreeningTime
 import java.time.LocalTime
 
 class ScreeningTimeTest {
