+++ app/src/test/java/woowacourse/movie/domain/model/reservation/seat/SeatingChartTest.kt
@@ -1,9 +1,11 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.domain.model.reservation.seat
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
+import woowacourse.movie.domain.model.reservation.seat.SeatRank
+import woowacourse.movie.domain.model.reservation.seat.SeatingChart
 
 class SeatingChartTest {
     @ParameterizedTest
